% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcm.r
\name{add.paths.landscape}
\alias{add.paths.landscape}
\title{Plot the paths on the prototype landscape}
\usage{
\method{add.paths}{landscape}(x, ...)
}
\arguments{
\item{x}{a tcm object}

\item{...}{Arguments to be passed to arrow(s)}
}
\description{
Plot the paths on the prototype landscape
}
