% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcm.r
\name{add.paths}
\alias{add.paths}
\title{add.paths}
\usage{
add.paths(x, ...)
}
\arguments{
\item{x}{a tcm object}

\item{...}{additional arguments}
}
\description{
A generic function for adding developmental trajectory onto the 
the prototype landscape
}
\author{
Wuming Gong, \email{gongx030@umn.edu}
}
