% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcm.r
\name{as.igraph.landscape}
\alias{as.igraph.landscape}
\title{Get an igraph object of neighboring graph of the prototype landscape where each vertex represent a prototype
and the distance between neighboring vertex is the Euclidean distance of their metagene coefficients}
\usage{
\method{as.igraph}{landscape}(x, ...)
}
\arguments{
\item{x}{a tcm object}
}
\description{
Get an igraph object of neighboring graph of the prototype landscape where each vertex represent a prototype
and the distance between neighboring vertex is the Euclidean distance of their metagene coefficients
}
