% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcm.r
\name{as.igraph}
\alias{as.igraph}
\title{as.igraph}
\usage{
as.igraph(x, ...)
}
\arguments{
\item{x}{an object}

\item{...}{additional arguments}
}
\description{
A generic function for converting prototype landscape to a igraph object
}
\author{
Wuming Gong, \email{gongx030@umn.edu}
}
