% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcm.r
\name{ctm}
\alias{ctm}
\title{The correlated topic model (CTM)}
\usage{
ctm(X, K, U = NULL, V = NULL, c0 = 0.001, d0 = 0.001,
  threshold = 0.001, max.iter = 200, verbose = TRUE)
}
\description{
The correlated topic model (CTM)
}
