% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcm.r
\name{fastmds}
\alias{fastmds}
\title{fastmds}
\usage{
fastmds(X, K, f = NULL, scale = FALSE, control = NULL)
}
\arguments{
\item{X}{The input gene by cell matrix}

\item{K}{Number of dimensions}

\item{f}{a vector indicating the pre-defined groups of input data}

\item{scale}{whether or not rescale the initialized metagene coefficients (default: FALSE)}

\item{control}{control parameters}
}
\description{
A sampling-based fastMDS algorithm based on Yang et al.
}
