% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcm.r
\name{gtm.core}
\alias{gtm.core}
\title{The generative topographic model}
\usage{
gtm.core(V, landscape = NULL, CT = NULL, mem = NULL,
  update.beta = FALSE, control = NULL)
}
\arguments{
\item{V}{a K by M metagene coefficient of cells, where K is the number of metagenes and M is the number of cells}
}
\description{
The generative topographic model
}
