% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcm.r
\name{gtm}
\alias{gtm}
\title{The generative topograhic map}
\usage{
gtm(V, Theta = NULL, L.scaffold = NULL, L.metacell = NULL, C = NULL,
  S = NULL, W = NULL, beta = 1, update.beta = TRUE, Alpha0 = 0.001,
  max.iter = 50, verbose = TRUE)
}
\description{
The generative topograhic map
}
