% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcm.r
\name{init.Theta.20171116b}
\alias{init.Theta.20171116b}
\title{Initialize metacell landscape}
\usage{
init.Theta.20171116b(V, CT, TMC, n.circle, n.metacell, is.scaffold, S.conv, L0,
  max.iter = 100)
}
\description{
Initialize metacell landscape
}
