% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcm.r
\name{init.Theta}
\alias{init.Theta}
\title{Initialize Theta for metacell landscape}
\usage{
init.Theta(method = 1, V, CT, TMC, n.circle, n.metacell, is.scaffold, S,
  S.conv, L, L0, Y, max.iter = 100, ...)
}
\description{
Initialize Theta for metacell landscape
}
