% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcm.r
\name{jitter2}
\alias{jitter2}
\title{Adding an Enclidean noise, x must be in a two dimensional matrix}
\usage{
jitter2(x, amount = 0.1)
}
\description{
Adding an Enclidean noise, x must be in a two dimensional matrix
}
