% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcm.r
\name{landscape}
\alias{landscape}
\title{landscape}
\usage{
landscape(type = "temporal.convolving", K = 15, ...)
}
\arguments{
\item{type}{type of the prototype landscape (default: temporal.convolving)}

\item{n.prototype}{prototypes per layer}

\item{n.circle}{the number of layer}

\item{n.prev}{number of convolving layers per time point}
}
\value{
a landscape object
}
\description{
Define the prototype landscape for simulating and visualizing scRNA-seq data
}
\examples{
# a temporal convolving landscape with the following parameters:
#  number of metagenes (K): 15
#  number of prototypes per layer (n.prototype): 15
#  number of layers per time point (n.circle): 10
#  number of convolving layers per time point (n.prev): 3
ls <- landscape(type = 'temporal.convolving', K = 15, n.prototype = 15, n.circle = 10, n.prev = 3)

# a plate landscape with following parameters:
#  number of metagenes (K): 15
#  number of prototypes per layer (n.prototype): 10
#  number of layers (n.circle): 100
ls <- landscape(type = 'plate', K = 15, n.prototype = 10, n.circle = 100)

}
\author{
Wuming Gong, \email{gongx030@umn.edu}
}
