% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcm.r
\name{ldpois}
\alias{ldpois}
\title{log density of Poission distribution
(faster than dpois(..., log = TRUE) due to the matrix operations)}
\usage{
ldpois(k, mu)
}
\description{
log density of Poission distribution
(faster than dpois(..., log = TRUE) due to the matrix operations)
}
