% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcm.r
\name{learning.rate}
\alias{learning.rate}
\title{learning rate for stochastic variational inference}
\usage{
learning.rate(t, control)
}
\arguments{
\item{t}{current iteration step}

\item{decay.rate}{down-weights for early iterations (default: 5)}

\item{forgetting.rate}{controls how quickly old information is forgotten (default: 0.75)}
}
\description{
learning rate for stochastic variational inference
}
