% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcm.r
\name{num2color}
\alias{num2color}
\title{Scale a continuous vector into color representation}
\usage{
num2color(x, cp = colorpanel(100, low = "blue", mid = "black", high =
  "yellow"))
}
\description{
Scale a continuous vector into color representation
}
