% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcm.r
\name{plot.tcm}
\alias{plot.tcm}
\title{plot tcm object.}
\usage{
\method{plot}{tcm}(x, ...)
}
\arguments{
\item{x}{tcm object}

\item{...}{Further graphical parameters may also be supplied as arguments}
}
\description{
plot tcm object.
}
