% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcm.r
\name{pseudotime.table}
\alias{pseudotime.table}
\title{Get a time table that consider the pseodotime information}
\usage{
pseudotime.table(X, n = NA, time.table = NULL, K = 3,
  direction = "center", control = NULL)
}
\description{
Get a time table that consider the pseodotime information
}
