% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcm.r
\name{set.control}
\alias{set.control}
\title{Setting the default values for the arguments for optimizing TCM}
\usage{
set.control(control)
}
\arguments{
\item{max.size.dist.matrix}{The maximum size of the distance matrix (for fastmds)}
}
\description{
Setting the default values for the arguments for optimizing TCM
}
