% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcm.r
\name{sim.rnaseq.ts}
\alias{sim.rnaseq.ts}
\title{Simulating time series single cell RNA-seq data}
\usage{
sim.rnaseq.ts(N = 200, M = 50, K = 20, n.circle = 100,
  n.metacell = 10, lambda = 0.25, alpha0 = 0.1, library.size = c(1000,
  1e+05), n.neighbors = 7, n.lineage = 3, choose.leaves = "random",
  type = "sequential", n.time.points = 10)
}
\arguments{
\item{N}{number of genes}

\item{M}{number of cells}

\item{K}{number of metagenes}

\item{n.circle}{number of circles on the latent space}

\item{lambda}{the argument of a exponential decay model for introducing the dropout noise}

\item{alpha0}{the argument for a Dirichlet distribution for sampling metagene basis}

\item{library.size}{total number of reads per cell}

\item{n.neighbors}{number of neighbors when looking for nearby prototypes}

\item{n.lineage}{number of simulated lineages}

\item{type}{the type of differentiation models}

\item{n.time.points}{the number of simulated time points}

\item{n.circle}{prototypes per circle on the latent space}
}
\value{
A SummarizedExperiment object of simulated temporal scRNA-seq data
}
\description{
Simulating time series single cell RNA-seq data
}
\examples{
set.seed(1)
sim <- sim.rnaseq.ts(N = 2000, M = 500, n.lineage = 5, type = 'sequential', n.time.points = 5)

}
\author{
Wuming Gong, \email{gongx030@umn.edu}
}
