% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcm.r
\name{tcm.core}
\alias{tcm.core}
\title{The core function for optimizaing TCM}
\usage{
tcm.core(X, U = NULL, V = NULL, Theta = NULL, L.scaffold = NULL,
  L.metacell = NULL, C = NULL, S = NULL, W = NULL, beta = 1,
  update.beta = TRUE, c0 = 0.001, d0 = 0.001, Alpha0 = 0.001,
  max.iter = 100, verbose = TRUE)
}
\arguments{
\item{L.scaffold}{Precision matrix defined for the scaffold}

\item{L.metacell}{Precision matrix defined for the metacells}

\item{S}{a column stochastic matrix (colSums(S) == 1).  describing the relationship between scaffold and metacells.}

\item{W}{a binary matrix indicating the restricted area of cells on the metacell landscape; if W is NULL, any cells can be mapped to any metacells on the metacell landscape}
}
\description{
The core function for optimizaing TCM
}
