% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcm.r
\name{tcm}
\alias{tcm}
\title{Visualizing temporal scRNA-seq by topographic cell map}
\usage{
tcm(X, K = 10, time.table = NULL, n.circle = 10, n.metacell = 15,
  n.prev = 3, lambda = 1, remove.first.pc = FALSE, update.beta = TRUE,
  max.iter.ctm = NA, max.iter = 100, verbose = TRUE)
}
\arguments{
\item{X}{a normalized read count matrix where each row represents a gene and each column represents a cell}

\item{K}{the number of metagenes}

\item{time.table}{a cell by time point table indicating the source of each cell}

\item{n.circle}{prototypes per layer (S)}

\item{n.metacell}{the number of layers (R)}

\item{n.prev}{the number of layers of convolving prototypes (R - rho)}

\item{remove.first.pc}{whether or not removing the first principal component}
}
\value{
a tcm object
}
\description{
Visualizing temporal scRNA-seq by topographic cell map
}
\examples{
set.seed(1)
sim <- sim.rnaseq.ts(N = 2000, M = 500, n.lineage = 5, type = 'sequential', n.time.points = 5)
bg.lineage <- rainbow(sim$n.lineage)
bg.cell <- bg.lineage[sim$lineage]
time.table <- table(1:sim$M, factor(sim$time))
set.seed(1)
mf <- tcm(sim$X, time = time.table, n.circle = 10, n.metacell = 15, n.prev = 3, max.iter = 10)
set.seed(1)
dev.new(height = 10, width = 12)
par(mar = c(5, 5, 5, 15))
plot(mf, pch = 21, bg = bg.cell, cex = 2.25)
legend(par('usr')[2], par('usr')[4], 1:sim$n.lineage, bty = 'n', xpd = NA, pt.bg = bg.lineage, pch = 21, col = 'black', cex = 1.75)

}
\author{
Wuming Gong, \email{gongx030@umn.edu}
}
