% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcm.r
\name{update.localvar}
\alias{update.localvar}
\title{updating local variable E[V], E[exp(V)], a and E[Z]}
\usage{
update.localvar(X, U, V = NULL, V.exp = NULL, W = NULL,
  landscape = NULL, model = NA, control = NULL, ...)
}
\arguments{
\item{X}{a N by M expression matrix}

\item{U}{a N by K metagene basis matrix}

\item{V}{the initial values of a K by M metagene coefficient matrix}
}
\description{
updating local variable E[V], E[exp(V)], a and E[Z]
}
