% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcm.r
\name{sim.rnaseq.ts}
\alias{sim.rnaseq.ts}
\title{sim.rnaseq.ts}
\usage{
sim.rnaseq.ts(N = 2000, M = 500, ls, lambda.dropout = 0.25,
  alpha0 = 0.1, library.size = c(1000, 1e+05), n.lineage = 5,
  type = "sequential", n.time.points = 5, ...)
}
\arguments{
\item{N}{number of genes}

\item{M}{number of cells}

\item{ls}{the prototype landscape (default: a "plate" with 100 circles and the number of prototypes per circle of 10)}

\item{alpha0}{the argument for a Dirichlet distribution for sampling metagene basis (default: 0.1)}

\item{library.size}{total number of reads per cell (default: c(1e3, 1e5))}

\item{n.lineage}{number of simulated lineages (default: 5)}

\item{type}{the type of differentiation models (default: 'sequential')}

\item{n.time.points}{the number of simulated time points (default: 5)}

\item{...}{additional arguments}

\item{lambda}{the argument of a exponential decay model for introducing the dropout noise (default: 0.25)}
}
\value{
A SummarizedExperiment object of simulated temporal scRNA-seq data
}
\description{
Simulating time series single cell RNA-seq data
}
\examples{
# simulate a simple temporal scRNA-seq data with 2,000 genes, 500 cells and five different lineages.  
# The single cell data are sampled across five time points following a sequentail differentiation model. 
set.seed(122)
sim <- sim.rnaseq.ts(N = 2000, M = 500, n.lineage = 5, n.time.points = 5)

}
\seealso{
\code{\link{landscape}}
}
\author{
Wuming Gong, \email{gongx030@umn.edu}
}
