% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcm.r
\name{tcm.core}
\alias{tcm.core}
\title{The core function for optimizaing TCM}
\usage{
tcm.core(X, U = NULL, V = NULL, a = NULL, landscape = NULL, CT = NULL,
  mem = NULL, c0 = 0.001, d0 = 0.001, control = NULL)
}
\arguments{
\item{S}{a column stochastic matrix (colSums(S) == 1).  describing the relationship between free and convolving prototypes}

\item{L}{Precision matrix defined for the prototypes}

\item{optimization.method}{optimization method for TCM, either 'batch' or 'stochastic'}

\item{batch.size}{size of mini batch}

\item{mc.cores}{number of CPU cores (default: 1)}

\item{max.size.per.batch}{maximum samples per batch}
}
\description{
The core function for optimizaing TCM
}
