% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcm.r
\docType{package}
\name{tcm}
\alias{tcm}
\alias{tcm-package}
\alias{tcm}
\title{TCM: A package for visualizing temporal scRNA-seq data}
\usage{
tcm(X, time.table, ls, init = NULL, control = NULL)
}
\arguments{
\item{X}{a read count matrix where each row represents a gene and each column represents a cell}

\item{time.table}{a cell by time point table indicating the source of each cell}

\item{ls}{the prototype landscape for TCM 
(default: landscape(type = 'temporal.convolving', time.points = ncol(CT), K = 15, n.prototype = 15, n.circle = 10, n.prev = 3))}

\item{init:}{initialization parameters}

\item{control:}{control parameters}
}
\value{
a tcm object
}
\description{
TCM: A package for visualizing temporal scRNA-seq data

Visualizing temporal scRNA-seq by topographic cell map
}
\examples{

library(tcm)

# simulate a simple temporal scRNA-seq data with 2,000 genes, 500 cells and five different lineages.  
# The single cell data are sampled across five time points following a sequentail differentiation model. 
set.seed(122)
sim <- sim.rnaseq.ts(N = 2000, M = 500, n.lineage = 5, n.time.points = 5)
X <- assays(sim)$count
time.table <- colData(sim)$time.table
mf <- tcm(X, time.table = time.table)
bg.cell <- rainbow(5)[colData(sim)$lineage]
dev.new(height = 10, width = 12)
par(mar = c(5, 5, 5, 15))
plot(mf, pch = 21, bg = bg.cell, cex = 1.5)

legend(par('usr')[2], par('usr')[4], 1:sim$n.lineage, bty = 'n', xpd = NA, pt.bg = bg.lineage, pch = 21, col = 'black', cex = 1.75)

}
\author{
Wuming Gong, \email{gongx030@umn.edu}
}
